modulo/requirements.txt
absl-py==1.0.0
astunparse==1.6.3
av==9.0.0
black==22.1.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
docstring-to-markdown==0.10
flatbuffers==2.0
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.2
imutils==0.5.4
jedi==0.18.1
jedi-language-server==0.35.1
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==13.0.0
Markdown==3.3.6
mypy-extensions==0.4.3
numpy==1.22.2
oauthlib==3.2.0
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.64
opt-einsum==3.3.0
pandas==1.4.1
parso==0.8.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pygls==0.11.3
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tk==0.1.0
tkvideoplayer==1.3.0
tomli==2.0.1
typeguard==2.13.3
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
wrapt==1.13.3
zipp==3.7.0
